переименовал файлы в snake_case, сгруппировал методы для вложений и проектов
